% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.boxplot.2d.R
\name{alpha}
\alias{alpha}
\title{Alpha}
\usage{
alpha(color, alpha)
}
\arguments{
\item{color}{the name or idea of a R color}

\item{alpha}{a value in [0,1] defining the opacity wanted.}
}
\description{
A simple function to change the opacity of a color
}
