\name{maxtemp}
\alias{maxtemp}
\title{Daily maximum temperatures in Melbourne, Australia}
\description{Daily maximum temperatures in Melbourne, Australia, from 1981-1990. Leap days have been omitted.}
\format{Time series of frequency 365.}
\source{Hyndman, R.J., Bashtannyk, D.M. and Grunwald, G.K. (1996)
"Estimating and visualizing conditional densities". \emph{Journal of
Computational and Graphical Statistics}, \bold{5}, 315-336.}

\keyword{datasets}
\examples{plot(maxtemp)
}
