% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.boxplot.2d.R
\name{shades}
\alias{shades}
\title{Shades}
\usage{
shades(color, n)
}
\arguments{
\item{color}{the name or idea of a R color}

\item{n}{number or shades wanted}
}
\description{
A simple function to genarte shade of one color by changing its opacity
}
