% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.boxplot.2d.R
\name{shades}
\alias{shades}
\title{Shades}
\usage{
shades(color, n)
}
\arguments{
\item{color}{the name or ID of the R color use for the darkest shade}

\item{n}{number or shades wanted}
}
\description{
A simple function to generate colored shades by changing opacity of the original color
}
